CBAPI-2290: add easy way to add single approvals and blocks
Also added the create_from_usb_device method as a quick way to create
approvals from a USB device, and the approve() method on USBDevice that
creates and saves a new USBDeviceApproval all in one go.  The sample doc has
been updated to use the new approve() functionality and also show approval
deletion.